Extract materials helpers, add tests for makeOrderText
